Изменения в PaymentService: - Добавил в домен кастомные ошибки - В слое repository для redis и postgres исправил в некоторых местах типы с int на int64 - В слое repository для redis изменил обработку ошибки

+++ services/payment/internal/repository/postgres/payment_repo.go
@@ -14,8 +14,8 @@ import (
 )
 
 type paymentMethodDB struct {
-	ID         int     `db:"id"`
-	UserID     int     `db:"user_id"`
+	ID         int64   `db:"id"`
+	UserID     int64   `db:"user_id"`
 	ExternalID string  `db:"external_id"`
 	CardType   string  `db:"card_type"`
 	Last4      string  `db:"last4"`
